requirements.txt
# pip list
# pip show torch
# pip uninstall -r requirements.txt -y
# pip install -r requirements.txt
# https://pytorch.org/get-started/locally/#windows-installation

# Tự tải cái này
#pip install torch torchvision torchaudio --extra-index-url https://download.pytorch.org/whl/cu126
Pillow>=9.0.0    
scikit-image     
numpy>=1.21.0
pandas>=1.5.0
scikit-learn>=1.2.0
tensorboard
tqdm             
matplotlib              
tensorboard
peft
grad-cam 
opencv-python
streamlit
timm
psutil